requirements.txt
pyyaml==3.11
coverage==4.4.1
django-widget-tweaks==1.4.1
js2py==0.50